Fix docker setup with nonstandard gitconfigs
If the dev e.g. has a redirection from https to ssh in their global
gitconfig, this would stop bundler from installing gems from GitHub.

Also, the Gemfile.lock in the docker setup may currently not be empty,
it must initially be a copy of the one in the core. This system could be
changed in the future to use Gemfile.local and Gemfile.lock.local, like
the native Ruby devs do it.